refactor(uebermensch-pages): swap hono for astro on cloudflare (hybrid SSR/SSG)
All seven Hono routes move to Astro file-based pages on
@astrojs/cloudflare. Hybrid output:

- SSR for content: /, /reports/, /briefs/, /{reports,briefs,wiki}/[slug].
  R2 is continuously written, so content renders at request time and
  sets Cache-Control s-maxage=60/86400 + stale-while-revalidate so the
  Cloudflare edge absorbs most traffic without coupling deploys to
  `uber sync r2`.
- Prerendered at build: /404 and /robots.txt.

Rendering logic lives in pure functions under src/lib/ that take an
R2Bucket and return { status, title, bodyHtml }. Astro pages are thin
glue; tests exercise the lib functions directly instead of an HTTP
handler, so no Astro runtime is needed in the test suite.

wrangler.toml keeps the VAULT R2 binding and now points `main` at the
adapter output `./dist/_worker.js/index.js` plus an ASSETS directory.
`astro dev` wires the binding locally via platformProxy.

Verification:
- pnpm test  — 15/15 pass (8 markdown + 7 pages)
- pnpm typecheck (astro check) — 0 errors / 0 warnings
- pnpm build — succeeds; /404.html and /robots.txt prerendered
- pnpm dev on :8788 serves /, /reports/, /briefs/, /robots.txt and
  404s on missing slugs against an empty local R2
