+++ apps/uebermensch-pages/.gitignore
@@ -1,0 +1,2 @@
+.astro/
+dist/
