+++ apps/uebermensch-pages/astro.config.mjs
@@ -1,0 +1,11 @@
+import { defineConfig } from "astro/config"
+import cloudflare from "@astrojs/cloudflare"
+
+export default defineConfig({
+  output: "server",
+  adapter: cloudflare({
+    platformProxy: { enabled: true },
+  }),
+  compressHTML: true,
+  server: { port: 8788 },
+})
