+++ apps/uebermensch-pages/package.json
@@ -4,18 +4,22 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "dev": "wrangler dev",
-    "deploy": "wrangler deploy",
+    "dev": "astro dev",
+    "build": "astro build",
+    "preview": "astro build && wrangler dev",
+    "deploy": "astro build && wrangler deploy",
     "test": "vitest run",
     "test:watch": "vitest",
-    "typecheck": "tsc --noEmit"
+    "typecheck": "astro check"
   },
   "dependencies": {
+    "@astrojs/cloudflare": "^12.5.0",
+    "astro": "^5.7.0",
     "gray-matter": "^4.0.3",
-    "hono": "^4.6.0",
     "marked": "^14.0.0"
   },
   "devDependencies": {
+    "@astrojs/check": "^0.9.0",
     "@cloudflare/workers-types": "^4.20240925.0",
     "typescript": "^5.7",
     "vitest": "^3.0.0",
