+++ apps/uebermensch-pages/src/pages/404.astro
@@ -1,0 +1,8 @@
+---
+export const prerender = true
+import Shell from "../layouts/Shell.astro"
+---
+<Shell title="Not found">
+  <h1>Not found</h1>
+  <p>This page doesn't exist.</p>
+</Shell>
