+++ apps/uebermensch-pages/src/pages/index.astro
@@ -1,0 +1,10 @@
+---
+import Shell from "../layouts/Shell.astro"
+import { renderHome } from "../lib/pages.ts"
+
+const result = await renderHome(Astro.locals.runtime.env.VAULT)
+Astro.response.headers.set("Cache-Control", "public, s-maxage=60, stale-while-revalidate=300")
+---
+<Shell title={result.title}>
+  <Fragment set:html={result.bodyHtml} />
+</Shell>
